+++ Sources/UISNPhotos.m
@@ -191,7 +191,7 @@
     }
     if(!flag){
         if(self.selects.count >= self.maxCount){
-            UIAlertView *alert = [[UIAlertView alloc]initWithTitle:nil message:[NSString stringWithFormat:@"最多可选%ld张",self.maxCount] delegate:self cancelButtonTitle:@"确定" otherButtonTitles:nil, nil];
+            UIAlertView *alert = [[UIAlertView alloc]initWithTitle:nil message:[NSString stringWithFormat:@"最多可选%ld张",(long)self.maxCount] delegate:self cancelButtonTitle:@"确定" otherButtonTitles:nil, nil];
             [alert show];
             return;
         }
@@ -204,7 +204,7 @@
     }];
     
     
-    self.lbCount.text = [NSString stringWithFormat:@"已选(%ld)",self.selects.count];
+    self.lbCount.text = [NSString stringWithFormat:@"已选(%ld)",(long)self.selects.count];
     if(self.selects.count > 0){
         [self.btnConfirm setTitleColor:[UIColor whiteColor] forState:UIControlStateNormal];
         self.btnConfirm.enabled = TRUE;
@@ -369,7 +369,7 @@
         self.btnSelect.hidden = TRUE;
         self.lbCount.hidden = FALSE;
         
-        self.lbCount.text = [NSString stringWithFormat:@"%ld",index];
+        self.lbCount.text = [NSString stringWithFormat:@"%ld",(long)index];
     }else{
         self.btnSelect.hidden = FALSE;
         self.lbCount.hidden = TRUE;
